Limit Wagtail to <2.11
Wagtail 2.11 deprecates `wagtail.core.middleware.SiteMiddleware`.
Testing with Wagtail 2.11 led to failing tests due to this deprecation.

It is certainly desirable to support Wagtail version above 2.11 in the
future. This should be addressed by following the upgrade steps outlined
in the [release notes for 2.11](https://docs.wagtail.io/en/stable/releases/2.11.html#sitemiddleware-moved-to-wagtail-contrib-legacy)