Test moving page created redirect
Adding a test to aid development of the feature to create automatic
redirects for cases when a page is moved. This feature has previously
not been available, due to missing signals.

The `pre_page_move` and `post_page_move` signals are available since
Wagtail 2.10. Using these to implement the automatic redirect creation
is the goal of this branch.