Add redirect for page after move
When the `post_page_move` signal is received, the URL of the page before
the move is constructed and a redirect is created from that old URL to
the instance of the page that was moved.

The `post_page_move` signal was chosen over the `pre_page_move` signal,
because it delivers the moved page instance after the move.

The signal call receiver function does receive the `url_path_before`.
In the tests, this did contain a `/home/` section in the beginning of
the string. Because I am not sure where this is coming from or if this
is always the same value, I have chosen to construct the page url before
the move. The url is constructed from the parent page before the move
(which is also a kwarg to the receiver function) and the current slug of
the moved page.

Child pages of the moved page are currently not considered. This will
be handled in a next step.

See also [documentation for the `post_page_move` signal](https://docs.wagtail.io/en/stable/reference/signals.html#pre-page-move-and-post-page-move)