Test redirect create for child of moved page
Before, the creation of a redirect was only tested for the moved page
itself. If a page that is moved has descendants then the URLs of these
descendants also change.

A test is added to check that a redirect is also created for the child
of a page that is moved. This test is currently failing, because the
response to the child URL before the move is 404 and not the desired
301.