Create redirects for child pages of moved page
The necessary function already existed before to handle the child pages
of a page for which the slug is changed.

This recursive function is now called in the end of handling for a page
move that created the redirect for the moved page itself.

This implementation passes the previously created test.